refactor cli from controller code
- move bogrod class to controller.py
- keep cli code in __init__.py for now
